+++ src/pages/projects.tsx
@@ -6,7 +6,7 @@ import Row from "react-bootstrap/Row"
 import DefaultFooter from "../components/footer"
 import SEO from "../components/seo"
 import { usePages } from "../hooks/use-pages"
-import DefaultNavbar from "../components/navbar"
+import DefaultHeader from "../components/header"
 import ProjectCard from "../components/project"
 
 const ProjectsPage = (): JSX.Element => {
@@ -14,11 +14,10 @@ const ProjectsPage = (): JSX.Element => {
   return (
     <div>
       <SEO title="Proyectos" lang="es" />
-      <DefaultNavbar />
       <Container fluid="xl" className="ps-md-4 pe-md-4 pt-3 pb-3">
-        <h1 className="h3">Mis proyectos</h1>
+        <DefaultHeader />
+        <h2 className="h3">Todos mis proyectos</h2>
         <p>Estos son todos mis proyectos públicos</p>
-        <h2 className="h4">Últimos proyectos</h2>
         <Row>
           {projects.map(item => (
             <Col key={item.id} xs="12" className="p-md-1 pb-2">
