+++ src/styles.scss
@@ -66,11 +66,6 @@ a:active {
   border: 1px solid rgba(255, 255, 255,0.5);
 }
 
-section {
-  padding-top: 70px;
-  margin-top: -50px;
-}
-
 .img-200 {
   height: 200px;
 }
